Poll auctions to complete based on time elapsed instead of number of bids
This allows us to more handle auction clients more flexibly and anonymously - they don't have to register as a client, they just need to send a bid in. This also means we won't wait forever if a client drops off the network without unregistering.

This does mean we have no mechanism to end an auction early, but we could add this logic to auctionExpired() - a certain number of bids or even a "buy it now" price. More important auctions could have  a reduced bid duration.